Install Tobac

install_tobac() installs the Python modules needed to run Tobac functions. The packages are installed in a virtualenv, which by default is named "harp-py". This virtual environment is the default virtual environment in which all harp related Python modules are installed. If you already have the "harp-py" virtualenv, you will be asked if you want to remove the virtualenv prior to installation. Typically you would answer no to this.

install_tobac(envname = "harp-py", new_env = identical(envname, "harp-py"))

Arguments

envname

The virtual environment into which to install the tobac Python modules. Default is "harp-py".

new_env

Whether to start from a new environment by removing the envname if it already exists. The defulat is TRUE if envname already exists, and you will be prompted if you want to remove the virtualenv envname.